requirements.txt
lmdb
numpy
ninja
pillow
requests
tqdm
scipy
scikit-image
scikit-video
trimesh[easy]
configargparse
munch
wandb
sorcery
scikit-learn
kornia
omegaconf
matplotlib
six
opencv-python
ipdb
mmcv==1.7.0
facexlib # https://github.com/xinntao/facexlib
# pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@stable